Factored out instruction code, to make them more readable.
The branch function was already written this way; the added run_rtype_instruction and run_itype_instruction functions now do the same.
They take a lambda for the execution of the opcode, allowing most opcodes to be written in one line. This also makes very clear what operaction each opcode is doing.