Update README.md
Add some info about the emulator being WIP and not complete. I still need to update all of this info to reflect that in the future I will be doing parts of it off-stream, but as the format is currently, this is sufficient to solve some confusion that has happened earlier with the contents of this repo.